move performance impact indicator types within schedulable_query.ts interfaces file
